backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
google-genai>=1.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
slowapi>=0.1.9
pdfplumber>=0.11.0
fpdf2>=2.8.0
apscheduler>=3.10.0
